2 small utils, used to run modeler whith no external dependencies ( except JAXP ).
The code is taken from jakarta-tomcat-connectors, with minimal
modifications ( more for DomUtil ).

JMX doesn't use a lot of advanced introspection - and the current
implementation does that directly ( the only dep on beanutils is via
digester ). IntrospectionUtils is used mostly for Dom.

Note that a lot of code in IU is not used - but I preffer to keep
it in sync with j-t-c


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/modeler/trunk@139109 13f79535-47bb-0310-9956-ffa450edef68