Read mbean-descriptors using only DOM.
All my tests show the performance is the same or better - most likely
due to the fact that the memory overhead of DOM is compensated by
direct invocation ( instead of introspection ). It is also a small factor
( since we have the same overhead - we create our own tree ).

This needs to be used only when modeler is in the top loader and
we want minimal jars there.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/modeler/trunk@139111 13f79535-47bb-0310-9956-ffa450edef68