One of the biggest problems with JMX is the way the filter is specified. When a listener is registered - it can specify a filter, and the filter will be called before calling the listener.
The problem is - we have no way to optimize the calls. The filter is almost
useless - it's just one extra call that could be done in the listener
itslef.

FixedNotificationFilter is one way around the problem. If we detect a filter
is of this type - we'll know that we can find out what notifications will
be accepted when the listener is added, and avoid calling all filters for all
notifications.

If anyone knows of a better way - please let me know.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/modeler/trunk@139124 13f79535-47bb-0310-9956-ffa450edef68