Another round of changes / refactorings. I think things are getting closer to stable.
A number of bug fixes / enhancements in IntrospectionSource ( used if
no explicit metadata is found ).

If a descriptor is not found, we'll also look for mbeans-descriptors in
the same package with the class we are looking for.

Started to optimize a bit the notification helper - using the Fixed filter.

It is not perfect yet, but I think it's getting better.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/modeler/trunk@139139 13f79535-47bb-0310-9956-ffa450edef68