Some work on the base notification.
The goal is to make it recyclable and support "notes" ( i.e. int-based array accessors
to attributes - like "handles" or descriptors), that would make the jmx notification
a good replacement for the current callbacks in coyote and jk ( and anything else ).

This is not yet complete - we also need an mbean to act as an "Id manager" and
maintain the int ids.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/modeler/trunk@139156 13f79535-47bb-0310-9956-ffa450edef68