Few more changes ( to the new APIs ).
Eliminate the side-effects - the descriptor reader will return a list.

Improve the MbeansSource - it can read jboss or mlet style xml files
and create the mbeans. It no longer needs the type to be specified.
Also added a method to allow support for lifecycle methods on the
mbeans.

Most JMX applications include a lifecycle mechanism - and the method names
and behavior is pretty common. If an mbean is used in such environment -
the init/start/stop/destroy will be called automatically. The intention
is to allow for such mbeans to work if used outside such a container, and
to support this behavior in tomcat standalone.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/modeler/trunk@139159 13f79535-47bb-0310-9956-ffa450edef68