Few more fixes, and initial code to add support for persistence.
Removed some of the experiments on syntax and unimplemented features.
For classloader control it'll probably use regular mbeans.

The persistence will be implemented using the DOM tree - each mbean
name will be associated with the DOM Node. Updates will be propagated
from BaseModelMbean to the source - and they'll change the corresponding
<attribute> nodes. Then Save will just write the DOM - so all comments
and structure will be preserved.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/modeler/trunk@139180 13f79535-47bb-0310-9956-ffa450edef68