More testing, more fixes :-)
Use the thread loader when creating the param classes. We could do this
when we load the definition - but it would be teribly slow. We could
also save a class loader in the definition.

We really need to figure out how the class loading would work with
modeler. JMX has the nice concept of flat loader ( similar with
Jboss), but in many cases this won't work ( containers, etc ).


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/modeler/trunk@139215 13f79535-47bb-0310-9956-ffa450edef68